+++ hydra/core/state_migrations.py
@@ -4,7 +4,12 @@ import copy
 from collections.abc import Callable, Mapping
 
 from hydra.core.state_migration_calls import migrate_v6_to_v7
-from hydra.core.state_migration_calls_vk_parasite import migrate_v10_to_v11, migrate_v11_to_v12, migrate_v12_to_v13
+from hydra.core.state_migration_calls_vk_parasite import (
+    migrate_v10_to_v11,
+    migrate_v11_to_v12,
+    migrate_v12_to_v13,
+    migrate_v13_to_v14,
+)
 from hydra.core.state_migration_creator_consumers import migrate_v8_to_v9
 from hydra.core.state_migration_headless_creator import migrate_v7_to_v8
 from hydra.core.state_migration_hydrabox import migrate_v5_to_v6
@@ -108,6 +113,7 @@ MIGRATIONS: dict[int, Migration] = {
     10: migrate_v10_to_v11,
     11: migrate_v11_to_v12,
     12: migrate_v12_to_v13,
+    13: migrate_v13_to_v14,
 }